require-dev is not optional
Even if you run composer update with parameter --no-dev you have to change your require-dev section to the current contents. This is caused by the behavior of composer to check both sections, require as well as require-dev, even if it does not install the dev requirements. Composer tells this very obviously with the following message:

"Running update with --no-dev does not mean require-dev is ignored, it just means the packages will not be installed. If dev requirements are blocking the update you have to resolve those problems."

This messages appears and the update is not running if you only change the require section.