--no-dev is not used in the full update manual
In the first section "Aktualisierung des Shops" the --no-dev option is mentioned everytime it is possible to use and to update without the dev requirements that e.g. unwanted in a production environment. This is good but in the second section "Aktualisierung der Modulkonfigurationen" the option is not mentioned at all. If you follow the current instructions, it doesn't matter if you use --no-dev in the first section since you won't be using it in the second section. Then all the dev requirements will get installed.

Since composer require and composer remove trigger a composer update everytime they are executed and since both commands aren't capable of the --no-dev option, you have to stop the update with --no-update option and execute the composer update with option --no-dev afterwards. If you do so, the dev requirements won't be installed at all.