Update update-from-6.1.x-to-6.2.0.rst
If you set minimum-stability to "dev" in your composer.json and run

composer require --no-interaction oxid-esales/oxideshop-update-component

the dev-master of the update-component will be installed, which can't handle the update process (at the moment). You must use a stable version. For this reason, require the package with a version constraint:

composer require --no-interaction oxid-esales/oxideshop-update-component:"^1.0"

This results in a stable installation and therefore a successfull update.

please also check de