Changed file editing to command use
It's always recommended to use Composer commands instead of editing the composer.json manually.

Also the example wasn't selected wisely. This documentation part describes all update from 6.2.x to 6.2.y, but the example showed metapackage in version 6.2.0, which leads to the case you are on 6.1.x before. This use case is described in another section since it involves some more steps.