Added a hint + possible configuration idea
Some users had trouble configuring the rights & roles, since you need at least one role, that has an active right, with a minimum of one usergroup assigned to restrict the rights for others. It's kind of described in the docs, but may be still difficult to understand. This behavior also needs to some questions how to configure a system with many different roles. Therefore the example is made by adding a "full access" role + usergroup to blacklist every user in the first step and then whitelist specific usergroups with additonal roles. The hint should lead to a better understanding as well as an idea how to start.